# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: Café Run Club Rewards API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /auth/user:
    get:
      tags: [Auth]
      operationId: getCurrentAuthUser
      summary: Get the currently authenticated user
      parameters:
        - $ref: '#/components/parameters/AuthorizationSessionHeader'
        - $ref: '#/components/parameters/SessionCookie'
      responses:
        '200':
          description: Auth status resolved successfully.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AuthUserEnvelope'
  /login:
    get:
      tags: [Auth]
      operationId: beginBrowserLogin
      summary: Start the browser OIDC login flow
      parameters:
        - name: returnTo
          in: query
          required: false
          schema:
            type: string
      responses:
        '302':
          description: Redirect to the OIDC authorization endpoint.
  /callback:
    get:
      tags: [Auth]
      operationId: handleBrowserLoginCallback
      summary: Complete the browser OIDC login flow
      parameters:
        - name: code
          in: query
          required: false
          schema:
            type: string
        - name: state
          in: query
          required: false
          schema:
            type: string
        - name: iss
          in: query
          required: false
          schema:
            type: string
            format: uri
      responses:
        '302':
          description: Redirect to the returnTo path on success.
  /logout:
    get:
      tags: [Auth]
      operationId: logoutBrowserSession
      summary: Clear the session and begin OIDC logout
      parameters:
        - $ref: '#/components/parameters/AuthorizationSessionHeader'
        - $ref: '#/components/parameters/SessionCookie'
      responses:
        '302':
          description: Redirect to the OIDC provider logout URL.
...
  /strava/webhook:
    get:
      tags: [strava]
      operationId: verifyStravaWebhook
      summary: Verify Strava webhook subscription (hub.challenge)
      parameters:
        - name: hub.mode
          in: query
          required: true
          schema:
            type: string
        - name: hub.challenge
          in: query
          required: true
          schema:
            type: string
        - name: hub.verify_token
          in: query
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Challenge echoed back to Strava
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/StravaWebhookChallenge'
        '400':
          description: Bad request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorEnvelope'
        '403':
          description: Verify token mismatch
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorEnvelope'
    post:
      tags: [strava]
      operationId: handleStravaWebhookEvent
      summary: Receive Strava webhook activity events
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/StravaWebhookEvent'
      responses:
        '200':
          description: Event acknowledged
...
  /club-events:
    get:
      tags: [clubEvents]
      operationId: listClubEvents
      summary: List all club run events
      responses:
        '200':
          description: List of club events
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/ClubEvent'
        '401':
          description: Not authenticated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorEnvelope'
    post:
      tags: [clubEvents]
      operationId: createClubEvent
      summary: Create a club run event (admin only)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ClubEventInput'
      responses:
        '201':
          description: Created event
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ClubEvent'
        '400':
          description: Validation error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorEnvelope'
        '403':
          description: Forbidden
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorEnvelope'
  /club-events/{eventId}:
    patch:
      tags: [clubEvents]
      operationId: updateClubEvent
      summary: Update a club run event (admin only)
      parameters:
        - name: eventId
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ClubEventInput'
      responses:
        '200':
          description: Updated event
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ClubEvent'
        '403':
          description: Forbidden
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorEnvelope'
        '404':
          description: Not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorEnvelope'
    delete:
      tags: [clubEvents]
      operationId: deleteClubEvent
      summary: Delete a club run event (admin only)
      parameters:
        - name: eventId
          in: path
          required: true
          schema:
            type: integer
      responses:
        '200':
          description: Deleted
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/DeleteSuccess'
        '403':
          description: Forbidden
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorEnvelope'
        '404':
          description: Not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorEnvelope'
components:
  parameters:
    AuthorizationSessionHeader:
      name: Authorization
      in: header
      required: false
      description: Opaque session token — Bearer <sid>.
      schema:
        type: string
    SessionCookie:
      name: sid
      in: cookie
      required: false
      description: Browser session cookie.
      schema:
        type: string
  schemas:
...
    AuthUser:
      type: object
      required: [id, email, firstName, lastName, profileImageUrl]
      properties:
        id:
          type: string
        email:
          type: ["string", "null"]
          format: email
        firstName:
          type: ["string", "null"]
        lastName:
          type: ["string", "null"]
        profileImageUrl:
          type: ["string", "null"]
    AuthUserEnvelope:
      type: object
      required: [user]
      properties:
        user:
          oneOf:
            - $ref: '#/components/schemas/AuthUser'
            - type: 'null'
    ErrorEnvelope:
      type: object
      required: [error]
      properties:
        error:
          type: string
    DeleteSuccess:
      type: object
      required: [success]
      properties:
        success:
          type: boolean
...
    ClubEvent:
      type: object
      required: [id, name, date, createdAt]
      properties:
        id:
          type: integer
        name:
          type: string
        date:
          type: string
          format: date
        description:
          type: ["string", "null"]
        createdAt:
          type: string
          format: date-time
    ClubEventInput:
      type: object
      required: [name, date]
      properties:
        name:
          type: string
        date:
          type: string
          format: date
        description:
          type: string
...
    StravaWebhookChallenge:
      type: object
      required: ["hub.challenge"]
      properties:
        hub.challenge:
          type: string
    StravaWebhookEvent:
      type: object
      properties:
        object_type:
          type: string
          description: Type of object (e.g. "activity")
        aspect_type:
          type: string
          description: Type of change (e.g. "create", "update", "delete")
        object_id:
          type: integer
          description: The activity or athlete ID
        owner_id:
          type: integer
          description: The Strava athlete ID who owns the object
        subscription_id:
          type: integer
        event_time:
          type: integer
          description: Unix timestamp of the event
...